4.1.1

Approved-by: Michal Spanel <spanel@3dim-laboratory.cz>

+++ include/3dim/coremedi/data/CDicom.h
@@ -57,7 +57,7 @@ public:
     };
 
     //! Vector of dicom slice numbers.
-    typedef std::vector<long long> tDicomNumList;
+    typedef std::vector<double> tDicomNumList;
 
 public:
     //! Default constructor.
